feat(finance): extract receipts list into feature components (Phase 6a)
- Extract Receipts list logic into ReceiptsFeature/ReceiptsTable
- Extract table JSX into dedicated ReceiptRow component
- Enforce keyboard accessibility (role="button", onKeyDown) for table rows
