+++ features/receipts/index.ts
@@ -1,0 +1,2 @@
+export * from './components/receipts-table';
+export * from './components/receipt-row';
